--- API/req.txt
@@ -1,52 +1,0 @@
-annotated-types==0.7.0
-anyio==4.9.0
-certifi==2025.6.15
-charset-normalizer==3.4.2
-click==8.2.1
-colorama==0.4.6
-dnspython==2.7.0
-docopt==0.6.2
-email_validator==2.2.0
-fastapi==0.115.13
-fastapi-cli==0.0.7
-h11==0.16.0
-httpcore==1.0.9
-httptools==0.6.4
-httpx==0.28.1
-idna==3.10
-isodate==0.7.2
-itsdangerous==2.2.0
-Jinja2==3.1.6
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-mdurl==0.1.2
-numpy==2.3.1
-orjson==3.10.18
-pandas==2.3.0
-pydantic==2.11.7
-pydantic-extra-types==2.10.5
-pydantic-settings==2.10.1
-pydantic_core==2.33.2
-Pygments==2.19.2
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.1
-python-multipart==0.0.20
-pytz==2025.2
-PyYAML==6.0.2
-requests==2.32.4
-rich==14.0.0
-rich-toolkit==0.14.7
-shellingham==1.5.4
-six==1.17.0
-sniffio==1.3.1
-starlette==0.46.2
-typer==0.16.0
-typing-inspection==0.4.1
-typing_extensions==4.14.0
-tzdata==2025.2
-ujson==5.10.0
-urllib3==2.5.0
-uvicorn==0.34.3
-watchfiles==1.1.0
-websockets==15.0.1
-yarg==0.1.10
